+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/attachments/factory/VisualMediaAccessAsState.kt
@@ -16,33 +16,15 @@
 
 package io.getstream.chat.android.compose.ui.messages.attachments.factory
 
-import android.Manifest.permission.READ_EXTERNAL_STORAGE
-import android.Manifest.permission.READ_MEDIA_IMAGES
-import android.Manifest.permission.READ_MEDIA_VIDEO
-import android.Manifest.permission.READ_MEDIA_VISUAL_USER_SELECTED
 import android.content.Context
-import android.content.pm.PackageManager
-import android.os.Build
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
 import androidx.compose.runtime.produceState
-import androidx.core.content.ContextCompat
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.LifecycleOwner
-
-/**
- * Defines the possible states in which the visual media storage access can be.
- *
- * - [VisualMediaAccess.FULL] when the app has full access to the storage.
- * - [VisualMediaAccess.PARTIAL] when the app has partial access to the storage.
- * - [VisualMediaAccess.DENIED] when the app has no access to the storage.
- */
-internal enum class VisualMediaAccess {
-    FULL,
-    PARTIAL,
-    DENIED,
-}
+import io.getstream.chat.android.ui.common.permissions.VisualMediaAccess
+import io.getstream.chat.android.ui.common.permissions.resolveVisualMediaAccessState
 
 /**
  * Produces the current [VisualMediaAccess] as [State] that can be observed in a [Composable] function.
@@ -75,34 +57,5 @@ internal fun visualMediaAccessAsState(
         awaitDispose {
             lifecycleOwner.lifecycle.removeObserver(eventObserver)
         }
-    }
-}
-
-/**
- * Resolves the current [VisualMediaAccess] state based on the permissions granted by the user.
- *
- * @param context The context to use to check the permission grants.
- */
-internal fun resolveVisualMediaAccessState(context: Context): VisualMediaAccess {
-    val isPermissionGranted = { permission: String ->
-        ContextCompat.checkSelfPermission(context, permission) == PackageManager.PERMISSION_GRANTED
-    }
-    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU &&
-        isPermissionGranted(READ_MEDIA_IMAGES) &&
-        isPermissionGranted(READ_MEDIA_VIDEO)
-    ) {
-        // Full access on Android 13 (API level 33) or higher
-        VisualMediaAccess.FULL
-    } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE &&
-        isPermissionGranted(READ_MEDIA_VISUAL_USER_SELECTED)
-    ) {
-        // Partial access on Android 14 (API level 34) or higher
-        VisualMediaAccess.PARTIAL
-    } else if (isPermissionGranted(READ_EXTERNAL_STORAGE)) {
-        // Full access up to Android 12 (API level 32)
-        VisualMediaAccess.FULL
-    } else {
-        // Access denied
-        VisualMediaAccess.DENIED
     }
 }