+++ CHANGELOG.md
@@ -75,6 +75,7 @@
 ### 🐞 Fixed
 
 ### ⬆️ Improved
+- Add support for partial media access in `MediaAttachmentFragment` and `FileAttachmentFragment` for Android 14+. [#5524](https://github.com/GetStream/stream-chat-android/pull/5524)
 
 ### ✅ Added
 - Add `ThreadListView` component for showing the list of threads for the user. [#5491](https://github.com/GetStream/stream-chat-android/pull/5491)